Begin code for path selection, renaming
now that starting position selection is done, i need to design 2 paths for each position and be able to choose them on the brain screen. the paths are drawn on the field diagram. code from this commit was never tested, as i dont have a brain at home. also i renamed some variables to make it easier to understand.